+++ lib/mocks.dart
@@ -20,7 +20,31 @@ final sight1 = SightModel(
   lon: 45.71948,
   url:
       'https://lh5.googleusercontent.com/p/AF1QipMibb54mBg1-vl6_j83NIDudzFxI4SiF0EE13en=w408-h459-k-no',
-  details: '«Кофейное место» - варим ароматный кофе из свежеобжаренных зёрен!',
+  details: '«Кофейное место» - варим ароматный кофе из свежеобжаренных зёрен!'
+      'Новый виток в освоении космоса связан с именем 2-х летней дворняжки Лайки, '
+      'которая 3 ноября 1957-го совершила первый орбитальный полет вокруг Земли. '
+      'Эта тихая и очень ласковая собака стала также первым хвостатым космонавтом, чье имя было "рассекречено" и стало известно всему миру. '
+      'Однако долгое время о ее полете не сообщали всей правды, поскольку она была очень грустной. '
+      'Новый виток в освоении космоса связан с именем 2-х летней дворняжки Лайки, '
+      'которая 3 ноября 1957-го совершила первый орбитальный полет вокруг Земли. '
+      'Эта тихая и очень ласковая собака стала также первым хвостатым космонавтом, чье имя было "рассекречено" и стало известно всему миру. '
+      'Однако долгое время о ее полете не сообщали всей правды, поскольку она была очень грустной. '
+      'Новый виток в освоении космоса связан с именем 2-х летней дворняжки Лайки, '
+      'которая 3 ноября 1957-го совершила первый орбитальный полет вокруг Земли. '
+      'Эта тихая и очень ласковая собака стала также первым хвостатым космонавтом, чье имя было "рассекречено" и стало известно всему миру. '
+      'Однако долгое время о ее полете не сообщали всей правды, поскольку она была очень грустной. '
+      'Новый виток в освоении космоса связан с именем 2-х летней дворняжки Лайки, '
+      'которая 3 ноября 1957-го совершила первый орбитальный полет вокруг Земли. '
+      'Эта тихая и очень ласковая собака стала также первым хвостатым космонавтом, чье имя было "рассекречено" и стало известно всему миру. '
+      'Однако долгое время о ее полете не сообщали всей правды, поскольку она была очень грустной. '
+      'Новый виток в освоении космоса связан с именем 2-х летней дворняжки Лайки, '
+      'которая 3 ноября 1957-го совершила первый орбитальный полет вокруг Земли. '
+      'Эта тихая и очень ласковая собака стала также первым хвостатым космонавтом, чье имя было "рассекречено" и стало известно всему миру. '
+      'Однако долгое время о ее полете не сообщали всей правды, поскольку она была очень грустной. '
+      'Новый виток в освоении космоса связан с именем 2-х летней дворняжки Лайки, '
+      'которая 3 ноября 1957-го совершила первый орбитальный полет вокруг Земли. '
+      'Эта тихая и очень ласковая собака стала также первым хвостатым космонавтом, чье имя было "рассекречено" и стало известно всему миру. '
+      'Однако долгое время о ее полете не сообщали всей правды, поскольку она была очень грустной. ',
   type: AppStrings.cafe,
 );
 final sight2 = SightModel(
