+++ lib/ui/screen/sight_search/sight_search_screen.dart
@@ -7,7 +7,7 @@ import 'package:places/res/app_colors.dart';
 import 'package:places/res/app_dimensions.dart';
 import 'package:places/res/app_strings.dart';
 import 'package:places/res/app_typography.dart';
-import 'package:places/ui/screen/list/sight_details_screen.dart';
+import 'package:places/ui/sight_details/sight_details_screen.dart';
 import 'package:places/ui/screen/sight_search/search_settings.dart';
 import 'package:places/ui/screen/sight_search/widgets/history_list_tile.dart';
 import 'package:places/ui/screen/sight_search/widgets/search_list_tile.dart';
