+++ lib/ui/sight_details/sight_details_settings.dart
